#2 컴포넌트 ui 섹션 단위로 정리
