+++ src/components/Banner/Banner.tsx
@@ -1,0 +1,18 @@
+import { BannerContainer, BannerImg } from "./Banner.styles";
+
+function Banner() {
+  return (
+    <BannerContainer>
+      <BannerImg
+        src={
+          import.meta.env.BASE_URL + "assets/images/banner/banner_invite.png"
+        }
+      />
+      <BannerImg
+        src={import.meta.env.BASE_URL + "assets/images/banner/banner_brand.png"}
+      />
+    </BannerContainer>
+  );
+}
+
+export default Banner;
