+++ src/components/Section2/Section2.styles.ts
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+import "../../index.css";
 
 export const GoalContainer = styled.div`
   position: absolute;
