+++ src/components/Section3/Section3.styles.ts
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+import "../../index.css";
 
 export const StepContainer = styled.div`
   overflow: hidden;
