+++ src/components/Section4/Section4.styles.ts
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+import "../../index.css";
 
 export const ReasonContainer = styled.div`
   display: flex;
